backend/requirements.txt
# ─── Web Framework ───────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# ─── AI / Agents ─────────────────────────────────────────────
openai==1.30.1
langgraph==0.1.1
langchain==0.2.1

# ─── Database ────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ─── Cache / Queue ───────────────────────────────────────────
redis==5.0.4
celery==5.4.0

# ─── Data Fetching ───────────────────────────────────────────
httpx==0.27.0
requests==2.32.2
beautifulsoup4==4.12.3
lxml==5.2.2
yfinance==0.2.40

# ─── Data Processing ─────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
ta==0.11.0

# ─── WebSocket ───────────────────────────────────────────────
websockets==12.0

# ─── Utilities ───────────────────────────────────────────────
schedule==1.2.2
python-jose==3.3.0
passlib==1.7.4
